add test_index_abs_messages
